Additional auth checks
- Added allowed user list
- Additional tests
- Require auth for status requests too

+++ test/__init__.py
@@ -8,5 +8,5 @@ def setup():
     # Create __init__
 
 def teardown():
-    os.unlink('ssh.out')
-
+    if os.path.exists('ssh.out'):
+        os.unlink('ssh.out')
